# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/comment/list": {
            "get": {
                "summary": "get all comments",
                "operationId": "5d89d4193a297667ebe04c4495d94d07",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/like/list": {
            "get": {
                "summary": "get all like",
                "operationId": "255612dccd007ddb4f954596b735e4a0",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/post/list": {
            "get": {
                "summary": "get all posts",
                "operationId": "9a70912c4ec51c17f5f7b5d17d549fb5",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/me": {
            "get": {
                "summary": "Get logged-in user details",
                "operationId": "8120197dbf6b5752a156b9868aa8bd4e",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/list": {
            "get": {
                "summary": "Get logged-in user details",
                "operationId": "75be91004f8b3fd250a7e0640d72458b",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}